feat/safety-service complete, admin accounts can flag for deletion

+++ internal/service/fileService/srv.go
@@ -5,11 +5,13 @@ import (
 	"mime/multipart"
 
 	"tosinjs/cloud-backup/internal/entity/errorEntity"
+	"tosinjs/cloud-backup/internal/repository/fileRepo"
 	"tosinjs/cloud-backup/internal/service/awsService"
 )
 
 type fileService struct {
 	awsSVC awsService.AWSService
+	repo   fileRepo.FileRepository
 }
 
 type FileService interface {
@@ -17,13 +19,33 @@ type FileService interface {
 	DeleteFile(username, folderName, filename string) *errorEntity.ServiceError
 	GetFile(username, folderName, filename string) ([]byte, *errorEntity.ServiceError)
 	DeleteFolder(username, folderName string) *errorEntity.ServiceError
+	FlagFile(filename string) *errorEntity.ServiceError
 	ListFilesInFolder(username, folderName string) ([]string, *errorEntity.ServiceError)
 }
 
-func New(awsSVC awsService.AWSService) FileService {
+func New(awsSVC awsService.AWSService, repo fileRepo.FileRepository) FileService {
 	return fileService{
 		awsSVC: awsSVC,
+		repo:   repo,
 	}
+}
+
+func (f fileService) FlagFile(filename string) *errorEntity.ServiceError {
+	flagCount, svcErr := f.repo.FlagFile(filename)
+	if svcErr != nil {
+		return svcErr
+	}
+	if flagCount > 2 {
+		svcErr = f.repo.DeleteFile(filename)
+		if svcErr != nil {
+			return svcErr
+		}
+		svcErr = f.awsSVC.DeleteFile(filename)
+		if svcErr != nil {
+			return svcErr
+		}
+	}
+	return nil
 }
 
 func (f fileService) UploadFile(hFile *multipart.FileHeader, username, folderName string) *errorEntity.ServiceError {
@@ -41,6 +63,10 @@ func (f fileService) UploadFile(hFile *multipart.FileHeader, username, folderNam
 		filename = fmt.Sprintf("%s/%s", username, filename)
 	}
 
+	if svcErr := f.repo.UploadFile(username, filename); svcErr != nil {
+		return svcErr
+	}
+
 	return f.awsSVC.UploadFile(file, filename)
 }
 
@@ -54,6 +80,10 @@ func (f fileService) DeleteFile(username, folderName, filename string) *errorEnt
 
 	if fileName != "" {
 		fileName = fmt.Sprintf("%s%s", fileName, filename)
+	}
+
+	if svcErr := f.repo.DeleteFile(fileName); svcErr != nil {
+		return svcErr
 	}
 
 	return f.awsSVC.DeleteFile(fileName)
